requirement.txt
fastapi
uvicorn
google-generativeai
ddgs
